mcp/requirements.txt
mcp-use
fastapi
uvicorn
python-dotenv
langchain-openai
langchain-anthropic
langchain-google-genai
certifi 